Trying to resolve a git issue
